We have logout

+++ src/actions/auth.js
@@ -1,4 +1,4 @@
-import { fetchWoToken } from '../helpers/fetch';
+import { fetchWoToken, fetchWToken } from '../helpers/fetch';
 import {types} from '../types/types'
 import Swal from 'sweetalert2';
 
@@ -7,10 +7,9 @@ export const startLogin = ( email, password ) => {
     return async(dispatch) => {
 
         const res = await fetchWoToken( 'auth', { email, password }, 'POST' );
-        const body = await res.json();
-        const {errors} = body;
+        const {username, uid, token, msg, status, errors} = await res.json();
 
-        if(res.status === 400) {
+        if(status === 400) {
             if(errors !== undefined) {
 
                 //Show validations for email input
@@ -19,30 +18,29 @@ export const startLogin = ( email, password ) => {
                 //Show validations for password input
                 if(errors.password !== undefined) { return Swal.fire('Something gone wrong', errors.password.msg, 'error');}
 
-            } else { Swal.fire('Something gone wrong', body.msg, 'error')}
+            } else { Swal.fire('Something gone wrong', msg, 'error')}
 
         } else {
-            //Swal.fire('Logged In','' ,'success') 
-            Swal.fire({ icon: 'success', text: 'Logged in',showConfirmButton: false})
-            localStorage.setItem('token', body.token)
-            localStorage.setItem('token-imit-date', new Date().getTime())
+            Swal.fire({ icon: 'success', text: 'Logged in',showConfirmButton: false, timer:1000})
+            localStorage.setItem('token', token);
+            localStorage.setItem('token-imit-date', new Date().getTime());
             dispatch( 
-                login( { mail: body.mail, uid: body.uid }) 
+                login( { username: username, uid: uid })
             );
         }
     }
 }
 
 
-export const startRegister = ( username, email, password ) => {
+export const startRegister = ( uname, email, password ) => {
     
     return async(dispatch) => {
 
-        const res = await fetchWoToken( 'auth/register', { username, email, password }, 'POST' );
-        const body = await res.json();
-        const {errors} = body;
+        const res = await fetchWoToken( 'auth/register', { uname, email, password }, 'POST' );
+        const {username, uid, token, msg, status, errors} = await res.json();
+        
 
-        if(res.status === 400) {
+        if(status === 400) {
             if(errors !== undefined) {
 
                 //Show validations for username input
@@ -54,23 +52,53 @@ export const startRegister = ( username, email, password ) => {
                 //Show validations for password input
                 if(errors.password !== undefined) { return Swal.fire('Something gone wrong', errors.password.msg, 'error');}
 
-            } else { Swal.fire('Something gone wrong', body.msg, 'error')}
+            } else { Swal.fire('Something gone wrong', msg, 'error')}
 
         } else {
-            //Swal.fire('Logged In','' ,'success') 
-            Swal.fire({ icon: 'success', text: 'Logged in',showConfirmButton: false})
-            localStorage.setItem('token', body.token)
+            localStorage.setItem('token', token)
             localStorage.setItem('token-imit-date', new Date().getTime())
             dispatch( 
-                login( { mail: body.mail, uid: body.uid }) 
+                login( { username: username, uid: uid }) 
             );
         }
     }
 }
+
+
+export const startChecking = () => {
+    
+    return async(dispatch) => {
+
+        const res = await fetchWToken( 'auth/renew');
+        const body = await res.json();
+
+        if(res.status === 400) { 
+            Swal.fire('Something gone wrong', body.msg, 'error');
+            dispatch(finishChecking());
+        } 
+        else {
+            localStorage.setItem('token', body.token);
+            localStorage.setItem('token-imit-date', new Date().getTime());
+            dispatch( login( { username: body.username, uid: body.uid }) );
+        }
+    }
+}
+
+const finishChecking = () => ({ type: types.authCheckingFinished });
+
+
+export const startLogout = () => {
+    return (dispatch) => {
+        localStorage.clear();
+        dispatch( logout() );
+    }
+};
 
 const login = (user) =>(
     {
         type: types.authLogin,
         payload:user
     }
-)
+);
+
+const logout = () => { return ({ type:types.authLogout }) }
